+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ## 0.19.0
 
 - **FIX**: `ShadInput` `readOnly` not updating.
+- **BREAKING CHANGE**: Rename `ShadSelect` and form fields `controller` to `popoverController`.
 
 ## 0.18.7